requirements.txt
# Core dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-multipart==0.0.6

# Infrastructure scanning
docker==7.0.0
paramiko==3.4.0
fabric==3.2.2
psutil==5.9.6
httpx==0.25.1

# LLM integration
litellm==1.11.1
openai==1.3.7
anthropic==0.7.7

# Data storage
sqlalchemy==2.0.23
alembic==1.13.0
databases==0.8.0
aiosqlite==0.19.0

# Documentation generation
jinja2==3.1.2
markdown==3.5.1
pymdown-extensions==10.5
pygments==2.17.2
beautifulsoup4==4.12.2

# Diagram generation
graphviz==0.20.1
diagrams==0.23.4
pydot==1.4.2

# PDF generation
weasyprint==60.1
reportlab==4.0.7

# Utilities
pyyaml==6.0.1
python-dateutil==2.8.2
croniter==2.0.1
click==8.1.7
rich==13.7.0
tabulate==0.9.0

# Monitoring and notifications
prometheus-client==0.19.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
pytest-cov==4.1.0
httpx==0.25.1

# Development
black==23.12.0
ruff==0.1.7
mypy==1.7.1
